# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateRoomBody {
  creatorName: string;
  durationMinutes: number;
}

export type RoomInfoStatus =
  (typeof RoomInfoStatus)[keyof typeof RoomInfoStatus];

export const RoomInfoStatus = {
  waiting: "waiting",
  running: "running",
  finished: "finished",
} as const;

export interface RoomInfo {
  code: string;
  creatorName: string;
  durationMinutes: number;
  status: RoomInfoStatus;
  participantCount: number;
}

export interface ErrorResponse {
  error: string;
}